Fix shared data access in raytracer, tidy up render loop
Render loop lost track of which blocks had been rendered due to a
synchronization issue. Specifically, std::vector<bool> is special-cased
as a bitfield and so is pretty much guaranteed not to be atomic unless
synchronized externally.

Also:

- Clean up types of a few variables
- Clean up openmp sharing type of variables (may result in better
  optimization)
- Replace shared rendered block count with an openmp reduction (results
  in fewer locks)

Fixes: lp:1608289
* https://bugs.launchpad.net/kicad/+bug/1608289

+++ 3d-viewer/3d_rendering/3d_render_raytracing/c3d_render_raytracing.h
@@ -40,6 +40,7 @@
 #include <plugins/3dapi/c3dmodel.h>
 
 #include <map>
+#include <cstddef>
 
 /// Vector of materials
 typedef std::vector< CBLINN_PHONG_MATERIAL > MODEL_MATERIALS;
@@ -110,7 +111,7 @@ private:
     unsigned long int m_stats_start_rendering_time;
 
     /// Save the number of blocks progress of the render
-    unsigned int m_nrBlocksRenderProgress;
+    size_t m_nrBlocksRenderProgress;
 
     CPOSTSHADER_SSAO m_postshader_ssao;
